requirements.txt
aiohttp>=3.9.0
anthropic>=0.40.0
certifi>=2024.0.0
pywhatkit>=5.4
python-dotenv>=1.0.0
